+++ example/CompilerPassFileFinderOptionsExample.php
@@ -7,7 +7,7 @@ require __DIR__.'/../vendor/autoload.php';
 use LDL\DependencyInjection\CompilerPass\Finder\Options\CompilerPassFileFinderOptions;
 use LDL\File\Helper\DirectoryHelper;
 
-echo "This example finds compiler passes and compiles them\n\n";
+echo "This example tests CompilerPassFileFinderOptions\n\n";
 
 $options = CompilerPassFileFinderOptions::fromArray([
     'directories' => [__DIR__.'/Build'],
@@ -17,7 +17,10 @@ dump($options->toArray());
 
 $dir = DirectoryHelper::getSysTempDir();
 $file = $options->write($dir->mkpath('compiler_pass_options.json'));
-$options = CompilerPassFileFinderOptions::fromJSONFile((string) $file);
+
+echo "\nRecreate options from JSON file:\n\n";
+
+$options = CompilerPassFileFinderOptions::fromJsonFile((string) $file);
 
 dump($options->toArray());